backend/reviewcollector/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
gunicorn==21.2.0
requests==2.31.0
firebase-admin==6.6.0
python-dotenv==1.0.1

